requirements.txt
# Dependencias básicas del framework
fastapi==0.110.0
uvicorn==0.27.1
gunicorn==23.0.0

# Plantillas y procesamiento de formularios
jinja2==3.1.6
markupsafe==2.1.5
python-multipart>=0.0.22

# Base de datos
aiosqlite==0.19.0
sqlalchemy==2.0.28
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Validación de datos
pydantic==2.6.3
pydantic-settings>=2.1.0

# Seguridad
# Reemplazando python-jose[cryptography] por cryptography directa para evitar vulnerabilidades de ecdsa
cryptography>=46.0.5
authlib>=1.3.1  # Actualizado para solucionar vulnerabilidad de confusión de algoritmos

# Manejo de imágenes
Pillow>=12.1.1

# WebSockets y comunicación asíncrona
websockets==12.0
aiofiles==23.2.1

# Integración con API de IA
google-generativeai>=0.8.0
python-dotenv==1.0.1

# Dependencias transitivas con problemas de seguridad
zipp==3.19.1  # Fijando versión para solucionar el problema de bucle infinito
importlib-metadata==6.8.0  # Versión compatible con zipp>=3.19.1

pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
pytest-mock==3.11.1
pytest-cov==4.1.0
